refactored chat component styles and types are in its own file
