+++ client/src/utils/chatTypes.ts
@@ -1,0 +1,11 @@
+export interface ChatUIProps {
+    sessionId: string;
+}
+
+export interface Message {
+    id: string;
+    text: string;
+    isUser: boolean;
+    timestamp: string;
+    avatar: string;
+}
